Move dummy filler to common.py
