+++ src/adcm_pytest_plugin/utils.py
@@ -413,35 +413,3 @@ class catch_failed(AbstractContextManager):
     def __exit__(self, exctype, excinst, exctb):
         if exctype is not None and issubclass(exctype, self._exctype):
             raise AssertionError(self._msg) from excinst
-
-
-@allure.title("Add dummy objects to ADCM")
-def add_dummy_objects_to_adcm(adcm_client):
-    """Add a bunch of dummy objects to ADCM"""
-    with allure.step("Create provider"):
-        provider_bundle = adcm_client.upload_from_fs(get_data_dir(__file__, "provider"))
-        provider = provider_bundle.provider_prototype().provider_create(name="Some provider")
-    with allure.step("Create cluster for the further import"):
-        cluster_to_import_bundle = adcm_client.upload_from_fs(get_data_dir(__file__, "cluster_to_import"))
-        cluster_to_import = cluster_to_import_bundle.cluster_prototype().cluster_create(name="Dummy cluster to import")
-    with allure.step("Create a cluster with service"):
-        cluster_bundle = adcm_client.upload_from_fs(get_data_dir(__file__, "cluster_with_service"))
-        cluster = cluster_bundle.cluster_prototype().cluster_create(name="Cluster with services")
-        cluster.bind(cluster_to_import)
-    with allure.step("Add services"):
-        service_first = cluster.service_add(name="Dummy service")
-        service_second = cluster.service_add(name="Second service")
-    with allure.step("Add hosts"):
-        host_first = provider.host_create(fqdn="host_in_cluster")
-        host_second = provider.host_create(fqdn="host_in_cluster_second")
-        cluster.host_add(host_first)
-        cluster.host_add(host_second)
-    with allure.step("Set components"):
-        cluster.hostcomponent_set(
-            (host_first, service_first.component(name="first")),
-            (host_second, service_first.component(name="second")),
-            (host_second, service_second.component(name="third")),
-        )
-    with allure.step("Run task"):
-        task = cluster.action(name="action_on_cluster").run()
-        task.wait()
